// frontend/src/app/api/model/updateModuleDto.ts
/**
 * CoreData API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DayOfWeek } from './dayOfWeek';


export interface UpdateModuleDto { 
    moduleCode?: string | null;
    title?: string | null;
    description?: string | null;
    credits?: number | null;
    level?: number | null;
    academicYear?: number | null;
    semesterOfStudy?: number | null;
    term?: string | null;
    isCore?: boolean;
    runsFrom?: string;
    runsTo?: string;
    scheduledDay?: DayOfWeek;
    scheduledStartLocal?: string;
    scheduledEndLocal?: string;
}
export namespace UpdateModuleDto {
}